♻️ refactor(refactor-novas-funcionalidades)

+++ backend/apps/films/views.py
@@ -13,12 +13,14 @@ class FilmViewSet(viewsets.ModelViewSet):
         old_instance = self.get_object()
         if 'image' in serializer.validated_data:
             if old_instance.image and old_instance.image != serializer.validated_data['image']:
-                if os.path.isfile(old_instance.image.path):
-                    os.remove(old_instance.image.path)
+                if old_instance.image.name != 'default.jpg' and old_instance.image.name.startswith('images/'):
+                    if os.path.isfile(old_instance.image.path):
+                        os.remove(old_instance.image.path)
         serializer.save()
 
     def perform_destroy(self, instance):
         if instance.image:
-            if os.path.isfile(instance.image.path):
-                os.remove(instance.image.path)
+            if instance.image.name != 'default.jpg' and instance.image.name.startswith('images/'):
+                if os.path.isfile(instance.image.path):
+                    os.remove(instance.image.path)
         instance.delete()
